[stack] prevent proc. reordering
x86 allows reads to by-pass reads (TSO model) which allows for the load of the .age element to happen after the the head is loaded.
This allows another thread to come in do it's operations in such a way that the ABA problem would be applicable & the new age+k > age
would be read by the original process.