[core] Determine out of headers sooner
Remember the first page examined whenever looping through header pages
while mapping. When we run out of avaliable headers, record this first
header page examined. All all iterations of that for each page loop
stop when there either are no more pages or when we hit the known
'first full' page. This means we only check a subset of the header pages.
Its possible although highly unlikely that this will actually be slower
but it should improve performance for memory intensive programs.